% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{print.fru}
\alias{print.fru}
\title{Print the fru object}
\usage{
\method{print}{fru}(x, ...)
}
\arguments{
\item{x}{Model to print.}

\item{...}{Ignored.}
}
\value{
Invisibly, the same object \code{x}.
}
\description{
Prints the basic information about the fitted model, in particular the OOB error estimated (if enabled previously).
}
